teacher_app/requirements.txt
altair==5.3.0
annotated-types==0.7.0
anyio==3.7.1
attrs==23.2.0
Authlib==1.3.0
azure-core==1.30.2
azure-storage-blob==12.20.0
azure-identity==1.17.1
azure-keyvault-secrets==4.8.0
bcrypt==4.1.3
blinker==1.8.2
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
distro==1.9.0
dnspython==2.6.1
extra-streamlit-components==0.1.71
streamlit-extras==0.4.3
Flask==3.0.2
fonttools==4.53.0
gitdb==4.0.11
GitPython==3.1.43
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mss==9.0.1
numpy==1.26.4
openai==1.2.4
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==5.27.2
pyarrow==16.1.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.8.0
pymongo==4.6.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.4.7
st-clickable-images==0.0.3
streamlit==1.36.0
streamlit-authenticator==0.2.3
tenacity==8.4.2
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
